Added logout manager class in manager layer
